switch solver to gameMechanics
